pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "reversible-anonymizer"
version = "1.0.13"
authors = [
    {name = "Omotayo Aina", email = "ainaomotayo@secureaiguard.com"},
]
description = "Enterprise-grade reversible anonymization using Google Cloud DLP"
readme = "README.md"
requires-python = ">=3.8"
license = {file = "LICENSE"}
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "Intended Audience :: Healthcare Industry",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Operating System :: OS Independent",
    "Topic :: Security",
    "Topic :: Security :: Cryptography",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "faker>=8.0.0",
    "google-cloud-dlp>=3.0.0",
    "google-cloud-firestore>=2.0.0",
    "google-cloud-service-usage>=1.0.0",
    "cryptography>=38.0.0",
    "python-memcached>=1.59",
    "google-cloud-memcache>=1.4.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=3.0.0",
    "black>=22.0.0",
    "isort>=5.10.0",
    "mypy>=0.950",
    "flake8>=4.0.0",
]
docs = [
    "sphinx>=4.0.0",
    "sphinx-rtd-theme>=1.0.0",
    "myst-parser>=0.18.0",
]

memcache = [
    "python-memcached>=1.59",
    "google-cloud-memcache>=1.4.1",
]

[project.urls]
"Homepage" = "https://github.com/ainaomotayo/reversible-anonymizer"
"Bug Tracker" = "https://github.com/ainaomotayo/reversible-anonymizer/issues"
"Documentation" = "https://reversible-anonymizer.readthedocs.io/"

[project.scripts]
anonymizer = "reversible_anonymizer.cli:main"

[tool.black]
line-length = 88
target-version = ["py38", "py39", "py310", "py311"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"